migrate ui components to bubbly
